fix: repair broken inject.html and merge fetch.js into converter.js
inject.html referenced JavaScript files left over from before the
switch to ES modules (namespace.js, converter.js, renderer.js) and
loaded inject.js as a classic script even though it uses import,
breaking the "Render selection" context menu action.

Also merge module/fetch.js into module/converter.js, its only caller.


# Conflicts:
#	changelog.adoc
