+++ app/html/inject.html
@@ -1,19 +1,13 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
 <head>
   <meta charset="utf-8"/>
-  <script type="text/javascript" src="../js/vendor/asciidoctor.js"></script>
   <script type="text/javascript" src="../js/vendor/highlight.js/highlight.min.js"></script>
-  <script type="text/javascript" src="../js/module/namespace.js"></script>
-  <script type="text/javascript" src="../js/module/settings.js"></script>
-  <script type="text/javascript" src="../js/module/dom.js"></script>
-  <script type="text/javascript" src="../js/module/theme.js"></script>
-  <script type="text/javascript" src="../js/converter.js"></script>
-  <script type="text/javascript" src="../js/renderer.js"></script>
+  <title>Asciidoctor.js Live Preview</title>
 </head>
 <body>
 <div id="content">
 </div>
-<script src="../js/inject.js" type="text/javascript"></script>
+<script src="../js/inject.js" type="module"></script>
 </body>
 </html>
